Update effect and example framework.
Effects now use a slightly more obvious syntax. Rather than `Kiwi.Plugins.ParticlePack3.AcidBullet( state, x, y )`, we now use `Kiwi.Plugins.ParticlePack3.AcidBullet.create( state, x, y )`. While slightly longer, this makes it obvious that the invocation is NOT using `new`. In addition, we now store the JSON for the particles in e.g. `AcidBullet.config`, rather than generating it anew every time we initialise a particle. This means it's easier to write new effects - they're just a standard `for` loop over the config array, so we can just copy-paste JSON and be done.

Example now uses a smart selector from the HTML. The dropdown option specifies the name of the effect, e.g. `AcidBullet`, which is immediately invoked with `create()`. This is a significant improvement on the prior method, which would create one of every effect and then add only the one that was selected.

A handful of effects have been defined to test this framework, and work perfectly.