Unify effect creation methods. Build docs.
As the `create` method on all the effects was identical, and always assumed effects were groups (not single-object effects), it made sense to shift everything to a single, more capable, method on the root object.

The new method also adds utility functions to the groups for `scheduleStop`, `startEmitting`, and `stopEmitting`.